Centralized reading, manipulating and writing configuration files
Introduce two pieces of middleware:

- store
- state

The first reads the configuration files, puts them on the request and
write them back to disks when they appear on the response.

The second accepts commands on the response which are used to
manipulates the configuration files and puts the new state on the
response to be dealt with by the store middleware.