Credentials moved to gateway.config.yml
The next version of the OOAPI gateway will not use a separate
credentials.json file, but keeps the credentails as configuration of
the gatewkeeper policy in the gateway.config.yml

This means we only need to deal with a single configuration file,
simplifying the store code.

This commit also introduces a ooapi-gateway-configurators.store.klist
namespace that helps to deal with the keyed-list constructs that are
used throughout that configuration file.