Fix select-all/none on institution with '.' in ID
The keyword-params skips params with a '.' character in it's name
causing this param to be passed as a string instead of a keyword.  This
mixing of strings and keywords yields an exception when sorted and may
cause other problems too.

This patchset disables keywordize-params and uses strings for
institution (and application) IDs in the store represenation to avoid
illegal keyword names.

Note: illegal keyword names are accepted by clojure for performance
reasons.  See also:

- https://web.archive.org/web/20210928081230/https://clojure.org/reference/reader#_literals
- https://web.archive.org/web/20210928081317/https://clojure.atlassian.net/browse/CLJ-17