Use datascript for datamodel
This patch replaces the `state` datastructure (a set of maps) by a
datascript database of entities. This should make it simpler to modify
entities and query the information necessary for views.

The assumption in this patch is that it's easy enough to work with the
datascript db directly most of the time, so the read-only db is passed
along to all the handlers. There are a few functions in the `model`
namespace for queries that are re-used and transactions that are
trickier. In particular, there is a generic method for dealing with
entities "belonging to" other entities (with the refs inverted from
datascript's native `:db/isComponent` attributes).