Notes field for institutions
Include a notes field for institutions to record information like
contact person, email address and telephone number for contacting the
person responsible for this endpoint.

It simply adds a "notes" key to a service endpoint in the yaml file.
Express gateway does not complain about the extra property and ignores
it.