Add and log domain events
This change adds an indirection mechnism where database changes are
reified as domain events (maps). The event maps are translated into
datascript transactions which are executed as earlier.

This means that all possible changes are visible in the
`model/event->tx` multi method, and allows for "domain level"
logging of changes.

The events are logged using the exising GELF logging mechanism, with
confidential information redacted.