Fix unable to login
The ring-defaults package switched from the in memory session store to
the cookie store.  Unfortunately this store is limited by the maximum
size of cookies (about 4k) which is not enough for the oauth session
data.  The formerly used memory session store should not be used for
production environments because it leaks memory.

This changeset introduces an in memory session store backed by
clojure.core.cache/ttl-cache-factory making sure old sessions are
purged.