Avoid creating a session when user not logged in yet
Session data is kept in memory (because cookies do not provide enough
space to keep the oauth stuff) so we should be careful when to allocate
memory for them.

This changeset ensure no is put on the session until the user has logged
in through OAUTH and thus a session is forced on them in the oauth2
middleware.